+++ pytest_django/plugin.py
@@ -361,6 +361,8 @@ def _django_db_marker(request):
             request.getfuncargvalue('transactional_db')
         else:
             request.getfuncargvalue('db')
+        if marker.serialized_rollback:
+            request.getfuncargvalue('serialized_rollback')
 
 
 @pytest.fixture(autouse=True, scope='class')
@@ -559,8 +561,9 @@ def validate_django_db(marker):
     It checks the signature and creates the `transaction` attribute on
     the marker which will have the correct value.
     """
-    def apifun(transaction=False):
+    def apifun(transaction=False, serialized_rollback=False):
         marker.transaction = transaction
+        marker.serialized_rollback = serialized_rollback
     apifun(*marker.args, **marker.kwargs)