+++ docs/changelog.rst
@@ -12,6 +12,12 @@ Bug fixes
 
 Features
 ^^^^^^^^
+* Add support for serialized rollback in transactional tests.
+  Thanks to Piotr Karkut for `the bug report
+  <https://github.com/pytest-dev/pytest-django/issues/329>`_.
+
+Features
+^^^^^^^^
 * Added a new option `--migrations` to negate a default usage of
   `--nomigrations`.